Compatibility with fedora?
It seems ansible-vault on Fedora doesn't play nice with stdin/stdout
using /dev/stdin and /dev/stdout.
I was using these explicitly because ansible-vault had a bug in 2.3.0.0
that broke using `-`:
https://github.com/ansible/ansible/issues/23567

But, they've fixed it and backported the fix, and Fedora does seem to be
able to handle it that way (and apparently--output doesn't matter anyway).